==> jupyter_py/3b-Load-IB-Data-Blocks.ipynb <==
## Import packages

**Remark:** ib_insync is a very useful third party library that can send *blocking* requests to Interactive Brokers. 

from IB_load_data import *
from datetime import datetime
from datetime import timedelta
from matplotlib import pyplot as plt
from ib_insync import *

import pandas as pd
import numpy as np

%load_ext autoreload
%autoreload 2

## Load Names of Tech Companies listed on NYSE and NASDAQ

nyse_tech_names = list(pd.read_csv(filepath_or_buffer = "../stock-names/nyse_tech_list.csv")["Symbol"])
nasdaq_tech_names = list(pd.read_csv(filepath_or_buffer = "../stock-names/nasdaq_tech_list.csv")["Symbol"])
nyse_energy_names = list(pd.read_csv(filepath_or_buffer = "../stock-names/nyse_energy_list.csv")["Symbol"])

## Connect to IB

ib = IB()
ib.connect('127.0.0.1', 7497, clientId=1)

## Load NYSE tech stocks

for i in range(len(nyse_energy_names)):
    stock_name = nyse_energy_names[i]
    
    print("Downloading data of NYSE: " + stock_name)
    
    contract = Stock(stock_name, 'SMART', 'USD', primaryExchange="NYSE")
    bars = ib.reqHistoricalData(contract, endDateTime="20190105 04:55:00", durationStr='10 Y',
            barSizeSetting='1 day', whatToShow='MIDPOINT', useRTH=True)
    
    # convert to pandas dataframe:
    df = util.df(bars)
    PATH_NAME = "../ib-data/nyse-daily-energy/" + str(stock_name) + ".csv"
    
    if df is None:
        print( "Error: " + str(i + 1) + "/" + str(len(nyse_energy_names)) + " download aborted.")
    else:
        df.to_csv(path_or_buf=PATH_NAME, index=False)
        print( str(i + 1) + "/" + str(len(nyse_energy_names)) + " completed.")

**Remarks**: 
- DELL is skipped because it is a newly relisted symbol on NYSE, no market data permissions to download 
- MAXR is skipped.

## Load NASDAQ tech stocks

- MIDPOINT: stopped at 228

for i in range(229, len(nasdaq_tech_names)):
    stock_name = nasdaq_tech_names[i]
    
    print("Downloading data of NASDAQ: " + stock_name)
    
    contract = Stock(stock_name, 'SMART', 'USD', primaryExchange="NASDAQ")
    bars = ib.reqHistoricalData(contract, endDateTime="20190105 04:55:00", durationStr='4 Y',
            barSizeSetting='5 mins', whatToShow='MIDPOINT', useRTH=True)
    
    # convert to pandas dataframe:
    df = util.df(bars)
    PATH_NAME = "../ib-data/iex/" + str(stock_name) + ".csv"

    if df is None:
        print( "Error: " + str(i + 1) + "/" + str(len(nasdaq_tech_names)) + " download aborted.")
    else:
        df.to_csv(path_or_buf=PATH_NAME, index=False)
        print( str(i + 1) + "/" + str(len(nasdaq_tech_names)) + " completed.")

contract = Stock('VXX', 'SMART', 'USD', primaryExchange="NYSE")
bars = ib.reqHistoricalData(contract, endDateTime="20190105 04:55:00", durationStr='4 Y',
        barSizeSetting='5 mins', whatToShow='MIDPOINT', useRTH=True)

# convert to pandas dataframe:
df = util.df(bars)
PATH_NAME = "../ib-data/nyse-vol/" + str('VXX') + ".csv"
df.to_csv(path_or_buf=PATH_NAME, index=False)
