==> jupyter_py/6a-Analyse-Pairs-Parameters.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt

def max_returns_per_lookback(summary):
    # returns the chart of the max returns given summary.csv
    summary_max = summary.groupby('lookback')['avg_overall_return'].max()

    X = summary_max.index
    Y = summary_max

    plt.plot(X,Y)
    plt.show()
    
def generate_mktcap_size(mkt_cap_info):
    
    # adds new column 'Size' based on market cap
    
    # mkt cap definition, link: https://www.nasdaq.com/article/market-cap-explained-cm954078
    # Mega-cap: Market value of $300 billion or more;
    # Large-cap: Market value between $10 billion and $300 billion;
    # Mid-cap: Market value between $2 billion and $10 billion;
    # Small-cap: Market value between $250 million and $2 billion;
    # Micro-cap: Market value of less than $250 million.

    mkt_cap_info['Size'] = 'Mega'
    mkt_cap_info.loc[mkt_cap_info['MarketCap'].between(0, 250000000, inclusive=False), 'Size'] = 'Micro'
    mkt_cap_info.loc[mkt_cap_info['MarketCap'].between(250000000, 2000000000, inclusive=False), 'Size'] = 'Small'
    mkt_cap_info.loc[mkt_cap_info['MarketCap'].between(2000000000, 10000000000, inclusive=False), 'Size'] = 'Mid'
    mkt_cap_info.loc[mkt_cap_info['MarketCap'].between(10000000000, 300000000000, inclusive=False), 'Size'] = 'Large'

    return mkt_cap_info

    
def get_pair_type(pair_data, mkt_cap_info):
    # given pair data and separate mkt_cap_info, this will return the SIZE-SIZE pairing as a new column
    pair_type = []
    for i in range(0,len(pair_data)):
        pair = [str(mkt_cap_info.loc[mkt_cap_info['Symbol'] == pair_data.pair.str.split("-")[i][0], "Size"].values[0]), str(mkt_cap_info.loc[mkt_cap_info['Symbol'] == pair_data.pair.str.split("-")[i][1], "Size"].values[0])]
        pair.sort()
        pair_type.append('-'.join(pair))

    pair_data['pair_type'] = pd.Series(pair_type) 

    return pair_data.groupby(['pair_type']).mean().sort_values(by = 'sharpe_ratio', ascending = False)



summary = pd.read_csv(filepath_or_buffer = "~/Desktop/summary.csv")
# summary = original_file[['lookback', 'enter_threshold_size', 'exit_threshold_size', 'loss_limit', 'avg_sharpe_ratio', 'avg_overall_return', 'overall_return_std']]

idx_lookback = summary.groupby(['lookback'])['avg_overall_return'].transform(max) == summary['avg_overall_return']
idx_enter_threshold = summary.groupby(['enter_threshold_size'])['avg_overall_return'].transform(max) == summary['avg_overall_return'] 
idx_exit_threshold = summary.groupby(['exit_threshold_size'])['avg_overall_return'].transform(max) == summary['avg_overall_return'] 
idx_loss_limit = summary.groupby(['loss_limit'])['avg_overall_return'].transform(max) == summary['avg_overall_return'] 


summary['avg_overall_return'] = round(summary['avg_overall_return'],4)*100
best_params = pd.concat([summary[idx_lookback], summary[idx_enter_threshold], summary[idx_exit_threshold], summary[idx_loss_limit]], join = 'inner')

# get the pair size info
nyse_info = pd.read_csv(filepath_or_buffer = "~/Desktop/FYP/statistical-arbitrage-private-18-19/stock-names/nyse_tech_list.csv")
mkt_cap_info = generate_mktcap_size(nyse_info[['Symbol', 'MarketCap']])                        
                                           

# summary parameters file by max overall returns
best_params.sort_values(by = 'avg_sharpe_ratio', ascending = False)[['lookback', 'enter_threshold_size', 'exit_threshold_size', 'loss_limit', 'avg_sharpe_ratio', 'avg_overall_return', 'overall_return_std']].drop_duplicates()

best_params.sort_values(by = 'avg_sharpe_ratio', ascending = False)

pair_data_good =  pd.read_csv(filepath_or_buffer = "~/Downloads/nyse-tech-daily/111434e3-d064-4989-ab85-d759b026671c.csv")
pair_data_good['returns_std_thousands'] = round(pair_data_good['returns_std']/1000,3)
get_pair_type(pair_data_good[['pair', 'sharpe_ratio', 'overall_return', 'returns_std_thousands']])


pair_data_med =  pd.read_csv(filepath_or_buffer = "~/Downloads/nyse-tech-daily/c8e80eda-b75e-46d6-9af9-2e5fea082f04.csv")
pair_data_med['returns_std_thousands'] = round(pair_data_med['returns_std']/1000,3)
get_pair_type(pair_data_med[['pair', 'sharpe_ratio', 'overall_return', 'returns_std_thousands']])

pair_data_bad =  pd.read_csv(filepath_or_buffer = "~/Downloads/nyse-tech-daily/89651f4b-6c52-4c8c-8f68-8279c63de971.csv")
pair_data_bad['returns_std_thousands'] = round(pair_data_bad['returns_std']/1000,3)
get_pair_type(pair_data_bad[['pair', 'sharpe_ratio', 'overall_return', 'returns_std_thousands']])

pair_data_good.sort_values(by = "overall_return", ascending = False)

borrow_cost =  pd.read_csv(filepath_or_buffer = "~/Desktop/borrow_cost.txt", sep = '|')
borrow_cost
